requirements/common.txt
#requrements/common.txt
#python==3.5.3
django==2.0.13
mysqlclient==1.3.14
celery==4.1.1
redis==3.2.0
requests==2.22.0
djangorestframework==3.11.0
simplejson
pycrypto==2.6.1
pyzabbix==0.7.5
uwsgi